api: fix issue which meant that float logged time endpoint was returning logged time after requested range
