+++ estimatey-api/Estimatey.Application/Common/Interfaces/IFloatClient.cs
@@ -7,4 +7,6 @@ public interface IFloatClient
     Task<List<LoggedTimeDto>> GetLoggedTime(int projectId, DateOnly? startDate = null, DateOnly? endDate = null);
 
     Task<List<FloatPersonDto>> GetPeople();
+
+    Task SyncPeople();
 }
